JavaScript Statements: Undefined Variables
You can use the JavaScript typeof operator to find the type of a JavaScript variable.
The typeof operator returns the type of a variable or an expression.

In JavaScript, a variable without a value, has the value undefined. The typeof is also undefined.

An empty value has nothing to do with undefined.
An empty string has both a legal value and a type.

In JavaScript null is "nothing". It is supposed to be something that doesn't exist.
Unfortunately, in JavaScript, the data type of null is an object.
You can consider it a bug in JavaScript that typeof null is an object. It should be null.
You can empty an object by setting it to null.